chore: restructure LabelLayoutUtil
